requirements.txt
Django==1.4b1
PIL==1.1.7
django-appconf==0.4.1
django-floppyforms==0.4.7
django-push==0.4
django-ratelimit-backend==0.1
django-staticfiles==1.1.2
feedparser==4.1
lxml==2.3.3
opml==0.5
psycopg2==2.4.1
pytz==2011n
wsgiref==0.1.2
